Added a parallel implementation of net train.
Use PatternParallelRectNet for parallelized training
of neural nets. PPRN creates a network and then creates a
thread pool for the children it will spawn.

Epoch starts.

PPRN creates PatternParallelNode children, which are really
just normal RectNets with a few small modifications. PPRN
then partitions the training data and hands it to the PPN
children, which train on that data and return the weight
changes that they incorporated at every weight in the
network.

PPRN waits for all of the PPN children to return their weight
changes, then incorporates these into its own net (scaling
down by the number of children that it spawned to begin
with).

Epoch ends.

At the end of this epoch, the PPRN checks to see if it needs
to continue training by computing the score across the entire
input set. If yes, another epoch begins. Else, the training
is terminated.